+++ src/main/java/com/poo/miapi/repository/core/TicketRepository.java
@@ -14,11 +14,11 @@ import java.util.Optional;
 
 @Repository
 public interface TicketRepository extends JpaRepository<Ticket, Integer> {
-  Optional<Ticket> findById(Long id);
+  Optional<Ticket> findById(int id);
 
   List<Ticket> findByEstado(EstadoTicket estado);
 
-  List<Ticket> findByCreadorId(Long idTrabajador);
+  List<Ticket> findByCreadorId(int idTrabajador);
 
   List<Ticket> findByTituloContainingIgnoreCase(String palabra);
 
@@ -44,8 +44,8 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
   List<Ticket> findByEstadoAndTecnicoActual(@Param("estado") EstadoTicket estado, @Param("tecnico") Tecnico tecnico);
 
   // Buscar tickets por estado y creador
-  List<Ticket> findByEstadoAndCreadorId(EstadoTicket estado, Long idTrabajador);
+  List<Ticket> findByEstadoAndCreadorId(EstadoTicket estado, int idTrabajador);
 
   // Métodos para estadísticas de SuperAdminService
-  long countByEstado(EstadoTicket estado);
+  int countByEstado(EstadoTicket estado);
 }
